+++ src/DateTimeImmutable.php
@@ -3,7 +3,7 @@
  * Date time immutable
  *
  * @author    Pronamic <info@pronamic.eu>
- * @copyright 2005-2021 Pronamic
+ * @copyright 2005-2022 Pronamic
  * @license   GPL-3.0-or-later
  * @package   Pronamic\WordPress\DateTime
  * @see       https://github.com/woocommerce/woocommerce/blob/3.3.4/includes/class-wc-datetime.php
@@ -21,29 +21,4 @@ namespace Pronamic\WordPress\DateTime;
  */
 class DateTimeImmutable extends \DateTimeImmutable implements DateTimeInterface {
 	use DateTimeTrait;
-
-	/**
-	 * Overrides upstream method to correct returned instance type to the inheriting one.
-	 *
-	 * {@inheritdoc}
-	 *
-	 * @param \DateTime $object Object.
-	 * @return self
-	 */
-	public static function createFromMutable( $object ) {
-		return self::create_from_mutable( $object );
-	}
-
-	/**
-	 * Create from mutable.
-	 *
-	 * @link https://www.php.net/manual/en/datetimeimmutable.createfrommutable.php
-	 * @param \DateTime $object The mutable DateTime object that you want to convert to an immutable version.
-	 * @return self
-	 */
-	public static function create_from_mutable( \DateTime $object ) {
-		$instance = new self( '@' . $object->getTimestamp() );
-
-		return $instance->setTimezone( $object->getTimezone() );
-	}
 }